requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.5
distro==1.9.0
docstring-parser==0.15
exceptiongroup==1.2.1
extracty==0.1.1
fake-useragent==1.5.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
instructor==0.6.8
jsonpatch==1.33
jsonpointer==2.4
langchain-community==0.0.29
langchain-core==0.1.52
langsmith==0.1.56
markdown-it-py==3.0.0
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.6
orjson==3.10.3
packaging==23.2
playwright==1.43.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
